[CHG] Tipo de dato

+++ account_move_reconcile/models/facturas_conciliacion.py
@@ -9,7 +9,7 @@ class FacturasConciliacion(models.Model):
     rzn_soc_emisor = fields.Char(string='Razón Social Emisor', size=128, index=True)
     rut_emisor = fields.Char(string='RUT Emisor', size=13, required=True, index=True)
     folio = fields.Integer(string='Folio', required=True, index=True)
-    monto_total = fields.Integer(string='Monto Total', required=True)
+    monto_total = fields.Float(string='Monto Total', required=True)
     fecha_emision = fields.Datetime(string='Fecha de Emisión', required=True)
     fecha_vencimiento = fields.Datetime(string='Fecha de Vencimiento')
     orden_compra = fields.Char(string='Orden de Compra', size=255, index=True)
